Add `setNativeProps` helper function on UI runtime (#4595)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently, there is no way to update props from a worklet imperatively
(except for calling JSI bindings `_updatePropsPaper` or
`_updatePropsFabric` directly, but they are implementation-detail and
meant to be private).

This PR adds `setNativeProps` helper function available on the UI
runtime which accepts an animated ref (created with `useAnimatedRef()`)
along with a plain JS object with props to be updated.

At the moment, both Paper and Fabric is supported as well as basic web
support works. Proper web support is doable but requires a refactor of
`useAnimatedRef`.

## Test plan

`SetNativePropsExample.tsx` 🆕