Update `useAnimatedGestureHandler` in docs (#4893)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Since `2.10.0` gesture handler enables new web implementation by
default. Because of that `useAnimatedGestureHandler` doesn't work unless
both `onGestureEvent` and `onHandlerStateChange` props are defined. This
PR updates docs entry for `useAnimatedGestureHandler` so that people
will be aware of that change.

---------

Co-authored-by: Michał Bert <michal.bert@swmansion.com>