Enable `NDEBUG` in production builds (#4902)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Related to #4901.

This PR properly sets `NDEBUG` flag when building in release mode.

References:
* https://github.com/facebook/react-native/pull/36172
* https://github.com/facebook/react-native/pull/38314
* https://github.com/facebook/react-native/pull/36194

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->