Add a `reduceMotion` prop to animation functions (#4748)
## Summary

Animations can cause some people to feel unwell. In order to help
developers create apps with those people in mind this PR adds a
reduceMotion prop to animation functions that allows developers to
configure whether a given animation should follow the system
accessibility setting. Possible values are `system`, `always` and
`never`. The prop defaults to `system` and is passed down to child
animations, unless they override it.

When reduced motion is enabled:
- `withSpring` and `withTiming` return the `toValue` immediately
- `withDecay` returns the current value immediately taking into account
the clamp parameter
- `withDelay` starts the next animation immediately
- `withRepeat`:
- If the numOfReps is infinite (<=0) or even and the animation is
reversed, then the repeated animation does not start and the startValue
is returned
    - otherwise the repeated animation is run once
- `withSequence` starts only the animations that have reduce motion
disabled (the last animation is also started, so the sequence ends on
the value it's supposed to)

This PR also adds changes to the Example App to reduce some animations
unrelated to reanimated.

## Test plan
Test the behavior of the ReducedMotion Example (and other examples) with
the accessibility setting on and off.