Add macOS support  (#4912)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds support for macOS and CI tests for building.

In order to make Reanimated work on macOS, we needed to:

- rename `ios/` directory to `apple/` directory - this makes native
modules visible both on iOS and macOS
- add `react-native-macos` as a dependency
- use `RCTUIView` and `RCTPlatformDisplayLink` on macOS instead of
`UIView` and `CADisplayLink` (which are still in use on iOS).


**Note:** At the moment we don't support several features on macOS,
including:
- Layout Animations
- Shared Element Transitions
- Sensors
- examples using libraries without macOS support, e.g.
[react-native-screens](https://github.com/software-mansion/react-native-screens)
or
[react-native-gesture-handler](https://github.com/software-mansion/react-native-gesture-handler)

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested using `MacOSExample`, as well as `Example/FabricExample` on
`iOS/Android`.


https://github.com/software-mansion/react-native-reanimated/assets/56112901/5f253794-82b4-4672-b7b3-eb352a9c8000