Add nightly CI to check TS compatibility with different RN versions (#4908)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Since types in `react-native-reanimated` depend on types from
`react-native` and we had some issues with type differences in e.g.
`0.71` and `0.72` we add this check to detect it early.

Requires #4899

## Test plan

🚀