Add navigation stack to macOS example app (#4975)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR added react-navigation/stack to MacOSExample app. As a result,
we are able to render an entire list of Reanimated examples, just like
on iOS or Android.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->