Add `window` availability check (#4976)
## Summary

When SSG is used to generate a website the `window` object can't be
accessed (in the server portion of the build). This PR adds an
availability check for when `window` is used.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->