Fix compatibility with older rn-screens version (#5014)
## Summary

Fixes
https://github.com/software-mansion/react-native-reanimated/issues/4992
This issue may happen when someone uses an older version of RN screens
without the `isModal` method. This PR adds a check to detect if this
selector exists. If not, the whole SET feature is disabled.

## Test plan

Try to compile the App with the newest Reaniamted and the old version of
rn-screens.