Don't warn against overwriting styles in case of layout animation (#5008)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Don't warn against overwriting styles in case of layout animation, since
most of use cases are valid

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

Tested with entering, layout and exiting animations
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>