Expose `isSharedValue` (#5019)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds `isSharedValue` function to the public API because
`@Shopify/react-native-skia` uses it:


https://github.com/Shopify/react-native-skia/blob/3563788df6fe0b75f48bc4300ab8a94f603a4662/package/src/external/reanimated/moduleWrapper.ts#L59

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->