Add reduce motion on macOS (#5032)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

This PR add support for Reduce Motion setting on macOS platform by
checking `accessibilityDisplayShouldReduceMotion` property from
`NSWorkspace`.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested with `Reduced Motion` example.