Add `Hover` to gesture names list in plugin (#5034)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR add auto-workletization of `Gesture.Hover` callbacks.

Reference:
https://docs.swmansion.com/react-native-gesture-handler/docs/next/api/gestures/hover-gesture/

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->