Fix worklet types [0/?] (#4794)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Since worklets are embedded deeply in the code (some might even say that
they are the backbone of reanimated❗).
I decided to split all the necessary changes to an unknown number of PRs
for gradual improvements and ease of review.

## Test plan

Newly introduced types in `src/reanimated2/commonTypes` should only be
present in the following files:
- `src/reanimated2/commonTypes.ts`,
- `src/reanimated2/threads.ts`.