Reduce file createAnimatedComponent - part 1 (#5053)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
File `createAnimatedComponent.tsx` contains our most important logic. It
is already very big and since in the future we want to support
${\textcolor{teal}{\textsf{inline Styles}}}$ this file may grow a lot.
Thats why we should make it clear and easy to rid - for example by
splitting it across many small files.

This refactor will be made in smaller parts.

### Part 1 ${\textcolor{green}{\textsf{(This PR)}}}$ 
* Create folder 📁 `createAnimatedComponent` 
* Move all files used only in `createAnimatedComponent.tsx` into that
folder. Move `createAnimatedComponent.tsx` as well.
* Extract all util function into separate file.
### Part 2
([LINK](https://github.com/software-mansion/react-native-reanimated/pull/5003))
* Create class `inlinePropUpdater.tsx` which handles logic related with
inline styles.
### Parts >2
This is not fixed yet, but I suggest we should:
* Move function `_filterNonAnimatedProps` away (70 lines of code)
  * Maybe create a separate class that would remember `initialStyle`? 
* Move away `setLocalRef` (inside `_setComponentRef`).
## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>