Add missing HostObject serialization in `makeShareableCloneOnUIRecursive` (#5048)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

#4617 Changed the logic of `makeShareableCloneOnUIRecursive`, detecting
if the object is a `HostObject` and skipping its serialization, on the
assumption that it's already good to go. This caused crashes in
`extractShareableOrThrow` due to how `isHostObject<T>` function works -
it not only checks if the object is a `HostObject` but also if it's of
type `T`.

<img width="631" alt="image"
src="https://github.com/software-mansion/react-native-reanimated/assets/40713406/64879db2-fdbb-4d8c-89f3-a37d97535d4f">

Therefore, given a `HostObject` from `react-native`, in this case
`ShadowNodeWrapper` on Fabric, it wasn't converted to our `Shareable`
type and failed the second check in `extractShareableOrThrow`.

Big thanks to @piaskowyk who debugged the most of it 🥳 

## Test plan

See that without the change in `makeShareableCloneOnUIRecursive` the new
error is thrown.

## Note

I have added some (seemingly) unnecessary `HostObject` creation elision,
please check if it's relevant to the problem or maybe should be deleted
or expanded.

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>