Fix Skia-Reanimated integration  (#5103)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes broken integration with Skia. The concept is to export
publicly `startMapper` and `stopMapper` for now but in the future
reexport those functions with different names so they would be more
explicit about what they do.

## Test plan

Check if Skia works.