Fix exporting enums as types and type of global in PlatformChecker (#5105)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixes regression that occurred in #5074.

Later on we also have to rethink if we want to promote `Extrapolate` or
`Extrapolation` since both are exactly the same and we only use the
former.

## Test plan

I actually made `Extrapolation` used in our examples so such mistake in
the future would be caught on CI (only regarding `Extrapolation` sadly).

> [!Note]
> I recommend using this
> ```json
> "editor.tokenColorCustomizations": {
>    "textMateRules": [
>      {
>        "scope": "entity.name.type.enum",
>        "settings": {
>          "foreground": "#a3e577"
>        }
>      },
>      {
>        "scope": "entity.name.type.class",
>        "settings": {
>          "foreground": "#e8b55e"
>        }
>      }
>    ]
>  }
>```
> In your VSCode settings so Classes, Enums and Types are not in the
same color.